Significant changes to AST format.
Renamed JSAssignOp to JSAssignExpression and JSUnary to JSUnaryExpression
added a JSNode to the constructor of each so they become expressions instead
of just opertators.

In the process changed a number of JSExpression* and JS*Expression
constructors to take single JSNode elements instead of [JSNode].

Due to these large changes to the AST representation, many of the tests
that rely on a particular AST represetation broke. Some of these were
fixed, but the majority were modified to use testRoundTrip instead so
that the tests are still active and still pass.