Lexer/Parser: Fix break/continue/return parsing
There were a number of issues around parsing these keywords when
when the code doesn't contain semi-colons. For example, this valid
code fragment would fail to parse:

     function f (x) {
         if (x) return
         x.foo()
     }

The solution was to modify the lexer so that it inserts an AutoSemiToken
after break/continue/return if and only if the white space immediately
following those keywords contain a newline.

Closes: https://github.com/erikd/language-javascript/issues/30
Closes: https://github.com/erikd/language-javascript/issues/36