AST: Improve representation of JSArrayLiterals
Array literals in Javascript have to include all sorts of rubbish like
elisions. The following are all examples of valid array literals:

    [], [,], [1], [,2], [3,], [,,4,,]

and so on. Handling this in the parser is relatively easy, but its not
easy to define a data type that allows these literals to be constructed
without also allow invalid literals like [1 2 3] (ie missing commas).