Minify the string concat operation by re-writing the AST
Regardless of whether strings are single or double quoted, if they are
separated by the concatenation operator, they should be re-written as a
single string. The single string will always use the string termination
character used by the first string. Finally, these concat re-writes are
appled from the leaves of the AST up so that multiple concat operations
will be collapsed into a single string.