Fix lexer bug involving template literals
In any program where a `}` character occurred anywhere before a template
literal, the lexer was incorrectly tokenizing the `}` and all characters
up through the initial backquote as a template tail. By adding an
additional piece of state to the parser monad, the lexer now will only
recognize a TemplateMiddleToken or a TemplateTailToken if the parser is
expecting one.