Fix minifier tests
The minifier was treating arrow functions with block bodies incorrectly
if they contained a single expression statement with no return
statement. It was transforming them into arrow functions with expression
bodies.

This means for instance that in
```js
let x = (() => { 42 })();
```

`x` would equal `undefined` before minification, but would equal `42`
after minification. The fix is to change the tests to match the correct
behavior.